+++ packages/modal/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @agencecinq/modal
 
+## 1.1.0
+
+### Minor Changes
+
+- Refacto
+
+## 1.0.1
+
+### Patch Changes
+
+- Update event system
+
 ## 1.0.0
 
 ### Major Changes